update print function

+++ monkey
@@ -1,12 +1,26 @@
 #!/bin/bash
 ###################################################################################################################################################################################################
+#通用打印函数
+OK='\e[0;32m'
+ERROR='\e[1;31m'
+END='\e[0m'
+ERROR()
+{
+	echo -e  "${ERROR} $1 ${END}"
+}
+
+OK()
+{
+	echo -e  "${OK} $1 ${END}"
+}
+###################################################################################################################################################################################################
 if [ $# != 1 ];then
-	echo "fatal: usage( ./do MR7063HC8W1 )"
+	ERROR "使用不正确，例如 (monkey MR7063HC8W1)"
 	exit 1
 fi
 
 if [ ! -e alps ];then
-	echo "Please return outside of alps floder exec!!!"
+	ERROR "执行路径不正确，请返回ALPS上级目录执行!!"
 	exit 1
 fi
 ###################################################################################################################################################################################################
@@ -17,9 +31,6 @@ CONFIGDIR=~/config
 SRCDIR=$PWD/alps
 CONFILE=$CONFIGDIR/config.ini
 PATCHDIR=$CONFIGDIR/patch
-OK='\e[0;32m'
-ERROR='\e[1;31m'
-END='\e[0m'
 
 cd $SRCDIR/system/vold/
 BRANCH=`git branch | awk '{if(match($1,"*")){print $2}}'`
@@ -36,15 +47,7 @@ DEFAULTXML=$SRCDIR/frameworks/base/packages/SettingsProvider/res/values/defaults
 BULIDFILE=$SRCDIR/build/target/product/core.mk
 SOUNDSDIR=$SRCDIR/frameworks/base/data/sounds
 
-ERROR()
-{
-	echo -e  "${ERROR} $1 ${END}"
-}
 
-OK()
-{
-	echo -e  "${OK} $1 ${END}"
-}
 
 ###################################################################################################################################################################################################
 #生成修改记录
@@ -77,7 +80,7 @@ echo "编译机型:$1\n" >> $RECORDFILE
 #修改机型名
 MODELNAME=`awk -F"=" '{if(/^机型名称/)print $2}' $CONFILE`
 if [ ! -z "$MODELNAME" ];then
-	OK ">>>>>Configurate Model Name = $MODELNAME"
+	OK ">>>>>修改机型名 = $MODELNAME"
 	sed -i "/^PRODUCT_MODEL/s/=.*/=$MODELNAME/" $PROFILE/elink_ID.mk
 	echo "修改机型名:$MODELNAME\n" >> $RECORDFILE
 fi
@@ -85,7 +88,7 @@ fi
 #修改蓝牙名称
 BLUETOOTHNAME=`awk -F"=" '{if(/^蓝牙名称/)print $2}' $CONFILE`
 if [ ! -z "$BLUETOOTHNAME" ];then
-	OK ">>>>>Configurate Bluetooth Name = $BLUETOOTHNAME "
+	OK ">>>>>修改蓝牙名称 = $BLUETOOTHNAME "
 	sed -i "/^bluetooth/s/=.*/=$BLUETOOTHNAME/" $CUSTOMCONF
 	echo "修改蓝牙名称:$BLUETOOTHNAME" >> $RECORDFILE
 fi
@@ -93,7 +96,7 @@ fi
 #修改Wifi共享热点SSID
 WLANSSID=`awk -F"=" '{if(/^共享SSID名称/)print $2}' $CONFILE`
 if [ ! -z "$WLANSSID" ];then
-	OK " >>>>>Configurate WLAN_SSID Display Label = $WLANSSID "
+	OK ">>>>>修改Wifi共享热点SSID = $WLANSSID "
 	sed -i "/^wlan.SSID/s/=.*/=$WLANSSID/" $CUSTOMCONF
 	echo "修改Wifi共享热点SSID:$WLANSSID" >> $RECORDFILE
 fi
@@ -101,7 +104,7 @@ fi
 #修改编译版本
 BUILDVERSION=`awk -F"=" '{if(/^编译版本/)print $2}' $CONFILE`
 if [ ! -z "$BUILDVERSION" ];then
-	OK " >>>>>Configurate Build version = $BUILDVERSION  "
+	OK ">>>>>修改编译版本 = $BUILDVERSION  "
 	sed -i "/^ELINK_VERSION/s/=.*/=$BUILDVERSION/" $PROFILE/elink_ID.mk
 	echo "修改编译版本:$BUILDVERSION" >> $RECORDFILE
 fi
@@ -109,7 +112,7 @@ fi
 #修改自定义编译版本
 CUSTOMBUILDVERSION=`awk -F"=" '{if(/^自定义编译版本/)print $2}' $CONFILE`
 if [ ! -z "$CUSTOMBUILDVERSION" ];then
-	OK " >>>>>Configurate Customer build version = $CUSTOMBUILDVERSION  "
+	OK ">>>>>修改自定义编译版本 = $CUSTOMBUILDVERSION  "
 	sed -i "/^CUSTOM_BUILD_VERNO/s/=.*/=$CUSTOMBUILDVERSION/" $SRCDIR/mediatek/config/common/ProjectConfig.mk
 	echo "修改自定义编译版本:$CUSTOMBUILDVERSION" >> $RECORDFILE
 fi
@@ -117,7 +120,7 @@ fi
 #修改时区
 TIMEZONE=`awk -F"=" 'gsub(/\//,"\\\/"){if(/^时区/)print $2}' $CONFILE`
 if [ ! -z "$TIMEZONE" ];then
-	OK " >>>>>Configurate Timezone = $TIMEZONE  "
+	OK ">>>>>修改时区 = $TIMEZONE  "
 	sed -i "/^persist.sys.timezone/s/=.*/=$TIMEZONE/" $PROFILE/system.prop
 	echo "修改时区:"`awk -F"=" '{if(/^时区/)print $2}' $CONFILE` >> $RECORDFILE
 fi
@@ -125,7 +128,7 @@ fi
 #修改默认亮度
 BRIGHTNESS=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^默认亮度/)print $2}' $CONFILE`
 if [ ! -z "$BRIGHTNESS" ];then
-	OK " >>>>>Configurate Screen brightness = $BRIGHTNESS  "
+	OK ">>>>>修改默认亮度 = $BRIGHTNESS  "
 	sed -i "/\"def_screen_brightness\"/s/>.*</>$BRIGHTNESS</" $DEFAULTXML
 	echo "修改默认亮度:$BRIGHTNESS" >> $RECORDFILE
 fi
@@ -133,7 +136,7 @@ fi
 #修改屏幕延时
 SCREENTIMEOUT=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^屏幕延时/)print $2}' $CONFILE`
 if [ ! -z "$SCREENTIMEOUT" ];then
-	OK " >>>>>Configurate screen timeout = $SCREENTIMEOUT  "
+	OK "修改屏幕延时 = $SCREENTIMEOUT  "
 	sed -i "/\"def_screen_off_timeout\"/s/>.*</>$SCREENTIMEOUT</" $DEFAULTXML
 	echo "修改屏幕延时:$(expr $SCREENTIMEOUT \/ 1000)秒" >> $RECORDFILE
 fi
@@ -141,7 +144,7 @@ fi
 #修改未知来源默认
 UNKNOWSRC=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^未知来源/)print $2}' $CONFILE`
 if [ ! -z "$UNKNOWSRC" ];then
-	OK " >>>>>Unkownsource selected = $UNKNOWSRC"
+	OK ">>>>>默认打开未知来源选项 = $UNKNOWSRC"
 	sed -i "/\"def_install_non_market_apps\"/s/>.*</>$UNKNOWSRC</" $DEFAULTXML
 	echo "默认打开未知来源选项" >> $RECORDFILE
 fi
@@ -149,7 +152,7 @@ fi
 #修改默认输入法
 INPUTMETHOD=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^默认输入法/)print $2}' $CONFILE`
 if [ ! -z "$INPUTMETHOD" ];then
-	OK " >>>>>Modify default input_method = $INPUTMETHOD  "
+	OK ">>>>>修改默认输入法 = $INPUTMETHOD  "
 	sed -i "/^DEFAULT_INPUT_METHOD/s/=.*/=$INPUTMETHOD/" $PROFILE/ProjectConfig.mk
 	echo "修改默认输入法:$INPUTMETHOD" >> $RECORDFILE
 fi
@@ -157,7 +160,7 @@ fi
 #修改可移动磁盘名
 DISKLABEL=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^可移动磁盘/)print $2}' $CONFILE`
 if [ ! -z "$DISKLABEL" ];then
-	OK " >>>>>Modify disk = $DISKLABEL "
+	OK ">>>>>修改可移动磁盘名 = $DISKLABEL "
 	cd $SRCDIR/system/core
 	
 
@@ -180,7 +183,7 @@ fi
 #修改联机ID
 ONLINELABEL=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^联机ID/)print $2}' $CONFILE`
 if [ ! -z "$ONLINELABEL" ];then
-	OK " >>>>>Modify online_id = $ONLINELABEL "
+	OK ">>>>>修改联机ID = $ONLINELABEL "
 	cd $SRCDIR/kernel/drivers
 
 	PRO=`expr substr $PROJECT 1 2`
@@ -200,7 +203,7 @@ fi
 #修改浏览器主页
 HOMEPAGE=`awk -F"=" 'gsub(/\//,"\\\/")sub(/^[[:blank:]]*/,"",$2){if(/^浏览器主页/)print $2}' $CONFILE`
 if [ ! -z "$HOMEPAGE" ];then
-	OK " >>>>>Modify default Browse Homepage = `expr substr $HOMEPAGE 10 20` "
+	OK ">>>>>修改浏览器主页 = `expr substr $HOMEPAGE 10 20` "
 	cd $SRCDIR/packages/apps/Browser
 	git apply --ignore-whitespace $PATCHDIR/homepage.patch
 	sed -i "/default homepage/s/,.*);/,\"$HOMEPAGE\");/" $SRCDIR/packages/apps/Browser/src/com/android/browser/BrowserSettings.java
@@ -270,7 +273,7 @@ if [ ! -z "$ACTIVEPROFILE" ];then
 		echo "情景模式输入错误，请按可供选择输入(标准，静音，会议，户外)"
 	fi
 
-	OK " >>>>>Modify Audio profile = $ACTIVEPROFILE  "
+	OK ">>>>>修改默认情景模式 = $ACTIVEPROFILE  "
 	sed -i "/\"def_active_profile\"/s/>.*</>$RESULT</" $SRCDIR/frameworks/base/packages/SettingsProvider/res/values/mtk_defaults.xml
 	echo "修改默认情景模式:$ACTIVEPROFILE" >> $RECORDFILE
 fi
@@ -287,12 +290,12 @@ SHUTTIMES=`echo $SHUTANIMATION | awk '{print $2}'`
 makeanimation()
 {
 	if [ $1 = "boot" ];then
-		OK " >>>>>begin to make bootanimation "
+		OK ">>>>>begin to make bootanimation "
 		cd $BOOTANIMATIONDIR
 		RESULT=bootanimation
 		TIMES=$BOOTTIMES
 	elif [ $1 = "shut" ];then
-		OK " >>>>>begin to make shutanimation "
+		OK ">>>>>begin to make shutanimation "
 		cd $SHUTANIMATIONDIR
 		RESULT=shutanimation
 		TIMES=$SHUTTIMES
@@ -307,13 +310,13 @@ makeanimation()
 	fi
 
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr -d ' '`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK " mv "$f" "$TARGET" "
+				OK "mv $f $TARGET "
 			fi
 		done
 	fi
@@ -331,7 +334,7 @@ makeanimation()
 		do 
 			INDEX=`expr $INDEX + 1` 
 			NAME=`printf "%04d\n" ${INDEX}`
-			OK " mv $i $RESULT/part0/${NAME}.$EXTENSION "
+			OK "mv $i $RESULT/part0/${NAME}.$EXTENSION "
 			mv $i $RESULT/part0/${NAME}.$EXTENSION
 			if [ "$EXTENSION" != "png" -a "$EXTENSION" != "PNG" ];then
 				convert $RESULT/part0/${NAME}.$EXTENSION $RESULT/part0/${NAME}.png
@@ -340,7 +343,7 @@ makeanimation()
 		done
 		mkdir -p $RESULT/part1
 		cp $RESULT/part0/${NAME}.* $RESULT/part1/
-		OK " mv $RESULT/part0/${NAME}.png $RESULT/part1/ "
+		OK "mv $RESULT/part0/${NAME}.png $RESULT/part1/ "
 	else
 		for i in ${FILES}
 		do 
@@ -352,7 +355,7 @@ makeanimation()
 				mkdir -p $RESULT/part$k
 			fi
 			NAME=`printf "%04d\n" ${INDEX}`
-			OK " mv $i $RESULT/part$k/${NAME}.$EXTENSION "
+			OK "mv $i $RESULT/part$k/${NAME}.$EXTENSION "
 			mv $i $RESULT/part$k/${NAME}.$EXTENSION
 	
 			if [ "$EXTENSION" != "png" -a "$EXTENSION" != "PNG" ];then
@@ -385,7 +388,7 @@ makeanimation()
 	fi
 	cp $RESULT.zip $SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJECT/system/media/
 	rm ../* -r
-	OK " make $RESULT successfully ===========> OK "
+	OK "make $RESULT successfully ===========> OK "
 	cd $CONFIGDIR
 }
 ###################################################################################################################################################################################################
@@ -409,10 +412,10 @@ HASKERNELLOGO=`ls $CONFIGDIR/custom/第二屏 | wc -l `
 makelogo()
 {
 	if [ $1 = "uboot" ];then
-		OK " >>>>>begin to make uboot logo "
+		OK ">>>>>begin to make uboot logo "
 		cd $UBOOTLOGODIR
 	elif [ $1 = "kernel" ];then
-		OK " >>>>>begin to make kernel logo "
+		OK ">>>>>begin to make kernel logo "
 		cd $KERNELLOGODIR
 	fi
 
@@ -421,13 +424,13 @@ makelogo()
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!!"
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr -d ' '`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK " mv 	"$f" "$TARGET" "
+				OK "mv $f $TARGET"
 			fi
 		done
 	fi
@@ -465,7 +468,9 @@ makelogo()
 			cp -p xganl_$1.bmp $SRCDIR/mediatek/custom/common/lk/logo/xganl/
 		fi
 	fi
-	rm *
+	if [ ! $1=="debug" ];then
+		rm *
+	fi
 	OK " make $1 logo  successfully ===========> OK "
 	cd $CONFIGDIR
 }
@@ -488,7 +493,6 @@ fi
 HASWALLPAPER=`ls $CONFIGDIR/custom/默认桌面壁纸 | wc -l `
 WALLPAPERDIR=$CONFIGDIR/custom/默认桌面壁纸
 if [ "$HASWALLPAPER" -gt 0  ];then
-	OK " >>>>>Change default wallpaper!! "
 	cd $WALLPAPERDIR
 
 	FILES=`ls`
@@ -496,13 +500,13 @@ if [ "$HASWALLPAPER" -gt 0  ];then
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr -d ' '`
 			if [ "$f" != "$TARGET" ];then
 					mv "$f" "$TARGET"
-					OK " mv "$f" "$TARGET" "
+					OK "mv $f $TARGET "
 			fi
 		done
 	fi
@@ -515,7 +519,7 @@ if [ "$HASWALLPAPER" -gt 0  ];then
 	fi
 	
 	rm * -r
-	OK " Change default  wallpaper  successfully ===========> OK "
+	OK "修改默认壁纸 ===========> OK "
 	cd $SRCDIR
 	echo "修改默认壁纸" >> $RECORDFILE
 fi
@@ -527,9 +531,7 @@ DATAAPPDIR=$SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJEC
 SYSTEMAPPDIR=$SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJECT/system/app
 BACKUPDIR=$SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJECT/system/appbackup
 if [ ! -z "$APKHANDLE" ];then
-	OK " >>>>>begin copy customer apk to android soruce!! "
 	cd $APKDIR
-
 	FILES=`ls`
 	NF=`ls -l |grep "^-"|wc -l`
 	FNF=`echo $FILES | awk '{print NF}'`
@@ -538,13 +540,13 @@ if [ ! -z "$APKHANDLE" ];then
 		return
 	fi
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr -d ' '`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK " mv "$f" "$TARGET" "
+				OK "mv $f $TARGET "
 			fi
 		done
 	fi
@@ -564,9 +566,9 @@ if [ ! -z "$APKHANDLE" ];then
 	if [ "$APKHANDLE" -eq "1" ];then
 		for i in `ls`
 		do
-			OK " copy $i to app "
+			OK "copy $i to app "
 			cp -p $i $DATAAPPDIR
-			OK " copy $i to appbackup "
+			OK "copy $i to appbackup "
 			cp -p $i $BACKUPDIR
 			echo "/data/app/$i" >> $DATAAPPDIR/.keep_list
 			echo "/system/appbackup/$i" >> $SYSTEMAPPDIR/.restore_list
@@ -578,7 +580,7 @@ if [ ! -z "$APKHANDLE" ];then
 	elif [ "$APKHANDLE" -eq "2" ];then
 		for i in `ls`
 		do
-			OK " copy $i to system_app "
+			OK "copy $i to system_app "
 			APKNAME=${i%\.*}
 			mkdir -p  $SRCDIR/vendor/common/SYSTEM_APP/$APKNAME
 			cp -p $i $SRCDIR/vendor/common/SYSTEM_APP/$APKNAME/
@@ -593,7 +595,7 @@ if [ ! -z "$APKHANDLE" ];then
 		echo "copy $i to app"
 		for i in `ls`
 		do
-			OK " copy $i to app "
+			OK "copy $i to app "
 			cp -p $i $SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJECT/data/app/
 			echo "/data/app/$i" >> $SRCDIR/vendor/mediatek/$PROJECT/artifacts/out/target/product/$PROJECT/data/app/.keep_list
 		done
@@ -601,17 +603,16 @@ if [ ! -z "$APKHANDLE" ];then
 		cd $CONFIGDIR
 		echo "预置APK(可卸载不可恢复)" >> $RECORDFILE
 	else
-		OK "${ERROE} ERROE:error apk handle!!!! "
+		ERROR "错误的选项，默认为(1,2,3)选项 "
 		return
 	fi
-
+	OK ">>>>>预置$NF个APK ===========> OK "
 fi
 ###################################################################################################################################################################################################
 #添加壁纸
 HASEXTRAWALLPAPER=`ls $CONFIGDIR/custom/备选壁纸 | wc -l `
 EXTRAWALLPAPERDIR=$CONFIGDIR/custom/备选壁纸
 if [ "$HASEXTRAWALLPAPER" -gt 0 ];then
-	OK " >>>>>begin copy extra wallpaper!! "
 	cd $EXTRAWALLPAPERDIR
 
 	FILES=`ls`
@@ -619,7 +620,7 @@ if [ "$HASEXTRAWALLPAPER" -gt 0 ];then
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr -d ' '`
@@ -655,13 +656,14 @@ if [ "$HASEXTRAWALLPAPER" -gt 0 ];then
 	done
 	rm * -r
 	cd $CONFIGDIR
+	OK ">>>>>预置壁纸$NF张 ===========> OK "
 	echo "预置壁纸:$NF张" >> $RECORDFILE
 fi
 ###################################################################################################################################################################################################
 #默认语言
 LANGUAGE=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^默认语言/)print $2}' $CONFILE`
 if [ ! -z "$LANGUAGE" ];then
-	OK " >>>>>default Language = $LANGUAGE "
+	OK ">>>>>修改默认语言 = $LANGUAGE "
 	sed -i "/DEFAULT_LATIN_IME_LANGUAGES/s/=.*/=$LANGUAGE/" $PROFILE/ProjectConfig.mk
 	sed -i "/MTK_PRODUCT_LOCALES/s/$LANGUAGE//" $PROFILE/ProjectConfig.mk
 	sed -i "/MTK_PRODUCT_LOCALES/s/=/=$LANGUAGE /" $PROFILE/ProjectConfig.mk
@@ -671,7 +673,7 @@ fi
 #开启ROOT权限
 OPENROOT=`awk -F"=" 'sub(/^[[:blank:]]*/,"",$2){if(/^开启ROOT权限/)print $2}' $CONFILE`
 if [ ! -z "$OPENROOT" ]; then
-	OK " >>>>>Open root permission "
+	OK ">>>>>开启ROOT权限 "
 	sed -i "/^EK_ROOT_SUPPORT/s/=.*/=$OPENROOT/" $COMMONPROFILE/ProjectConfig.mk
 	echo "开启ROOT权限" >> $RECORDFILE
 fi
@@ -680,7 +682,7 @@ fi
 HASDEFAULTRINGTONE=`ls $CONFIGDIR/custom/默认来电铃声 | wc -l `
 DEFAULTRINGTONEDIR=$CONFIGDIR/custom/默认来电铃声
 if [ "$HASDEFAULTRINGTONE" -gt 0 ]; then
-	OK " >>>>>Change default ringtone!! "
+	OK ">>>>>修改默认来电铃声 ===========> OK  "
 	cd $DEFAULTRINGTONEDIR
 
 	FILES=`ls`
@@ -688,13 +690,13 @@ if [ "$HASDEFAULTRINGTONE" -gt 0 ]; then
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		echo "Rename because any filename has blank!!!"
+		OK "Rename because any filename has blank!!!"
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr " " "_"`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK " mv "$f" "$TARGET" "
+				OK "mv $f $TARGET "
 			fi
 		done
 	fi
@@ -705,17 +707,15 @@ if [ "$HASDEFAULTRINGTONE" -gt 0 ]; then
 		sed -i '/^PRODUCT_COPY_FILES/a\	\$(LOCAL_PATH)/ringtones/'$RINGFILE':system/media/audio/ringtones/'$RINGFILE' \\' $SOUNDSDIR/AudioPackage2.mk
 		rm $DEFAULTRINGTONEDIR/$RINGFILE
 	else
-		echo  -e "${ERROR}  More than one audio file. "
+		ERROR "超过一个音频文件，最多只接受一个文件. "
 	fi
 fi
 ###################################################################################################################################################################################################
 #备选来电铃声
 HASEXTRARINGTONE=`ls $CONFIGDIR/custom/备选来电铃声 | wc -l `
 EXTRARINGTONEDIR=$CONFIGDIR/custom/备选来电铃声
-if [ "$HASEXTRARINGTONE" -gt 0 ]; then
-	OK " >>>>>Add extra ringtone!! "
+if [ "$HASEXTRARINGTONE" -gt 0 ]; then	
 	cd $EXTRARINGTONEDIR
-
 	FILES=`ls`
 	NF=`ls -l |grep "^-"|wc -l`
 	FNF=`echo $FILES | awk '{print NF}'`
@@ -727,7 +727,7 @@ if [ "$HASEXTRARINGTONE" -gt 0 ]; then
 			TARGET=`echo "$f" | tr " " "_"`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK "  mv "$f" "$TARGET" "
+				OK "mv $f $TARGET"
 				cp $EXTRARINGTONEDIR/$f $SOUNDSDIR/ringtones
 				sed -i '/^PRODUCT_COPY_FILES/a\	\$(LOCAL_PATH)/ringtones/'$f':system/media/audio/ringtones/'$f' \\' $SOUNDSDIR/AudioPackage2.mk
 			fi
@@ -740,21 +740,20 @@ if [ "$HASEXTRARINGTONE" -gt 0 ]; then
 		done
 	fi
 	rm $EXTRARINGTONEDIR/*
+	OK ">>>>>添加备选来电铃声 ===========> OK "
 fi
 ###################################################################################################################################################################################################
 #默认通知铃声
 HASDEFAULNOTIFICATIONE=`ls $CONFIGDIR/custom/默认通知铃声 | wc -l `
 DEFAULNOTIFICATIONDIR=$CONFIGDIR/custom/默认通知铃声
 if [ "$HASDEFAULNOTIFICATIONE" -gt 0 ]; then
-	OK " >>>>>Change default notifications!! "
 	cd $DEFAULNOTIFICATIONDIR
-
 	FILES=`ls`
 	NF=`ls -l |grep "^-"|wc -l`
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr " " "_"`
@@ -771,15 +770,15 @@ if [ "$HASDEFAULNOTIFICATIONE" -gt 0 ]; then
 		sed -i '/^PRODUCT_COPY_FILES/a\	\$(LOCAL_PATH)/notifications/'$NOTIFICATIONFILE':system/media/audio/notifications/'$NOTIFICATIONFILE' \\' $SOUNDSDIR/AudioPackage2.mk
 		rm $DEFAULNOTIFICATIONDIR/$NOTIFICATIONFILE
 	else
-		OK "${ERROR} ERROR More than one audio file.  "
+		ERROR "超过一个音频文件，最多只接受一个文件. "
 	fi
+	OK ">>>>>添加默认通知铃声 ===========> OK "
 fi
 ###################################################################################################################################################################################################
 #备选通知铃声
 HASEXTRANOTIFICATION=`ls $CONFIGDIR/custom/备选通知铃声 | wc -l `
 EXTRANOTIFICATIONDIR=$CONFIGDIR/custom/备选通知铃声
 if [ "$HASEXTRANOTIFICATION" -gt 0 ]; then
-	OK " >>>>>Add extra notifications!!  "
 	cd $EXTRANOTIFICATIONDIR
 
 	FILES=`ls`
@@ -787,13 +786,13 @@ if [ "$HASEXTRANOTIFICATION" -gt 0 ]; then
 	FNF=`echo $FILES | awk '{print NF}'`
 	
 	if [  $NF -ne $FNF ];then	
-		OK " Rename because any filename has blank!!! "
+		OK "Rename because any filename has blank!!! "
 		for f in `ls ./ | tr " " "\?"`
 		do
 			TARGET=`echo "$f" | tr " " "_"`
 			if [ "$f" != "$TARGET" ];then
 				mv "$f" "$TARGET"
-				OK " mv "$f" "$TARGET" "
+				OK "mv $f $TARGET"
 				cp $EXTRANOTIFICATIONDIR/$f $SOUNDSDIR/notifications
 				sed -i '/^PRODUCT_COPY_FILES/a\	\$(LOCAL_PATH)/notifications/'$f':system/media/audio/notifications/'$f' \\' $SOUNDSDIR/AudioPackage2.mk
 			fi
@@ -806,10 +805,13 @@ if [ "$HASEXTRANOTIFICATION" -gt 0 ]; then
 		done
 	fi
 	rm $EXTRANOTIFICATIONDIR/*
+
+	OK ">>>>>添加备选通知铃声 ===========> OK "
 fi
-
-
+###################################################################################################################################################################################################
 cd $CONFIGDIR
 OK "===============================>>out $RECORDFILE "
-git checkout -- config.ini
+if [ ! $1=="debug" ];then
+	git checkout -- config.ini
+fi
